Moved share credentials code from home to comfirm credentials and added list view to see credential types
