+++ QRCodeAuthMobile/QRCodeAuthMobile/QRCodeAuthMobile/ConfirmCredentials.xaml
@@ -3,24 +3,26 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d"
              x:Class="QRCodeAuthMobile.ConfirmCredentials">
     <ContentPage.Content>
         <StackLayout BackgroundColor="#005073" Padding="40">
 
-            <StackLayout BackgroundColor="Transparent" VerticalOptions="EndAndExpand">
+			<StackLayout BackgroundColor="Transparent" VerticalOptions="EndAndExpand">
 
-                <Label Text="Please Confirm Credential Sharing"
+				<Label Text="Please Confirm Credential Sharing"
                        HorizontalOptions="CenterAndExpand" 
                        Style="{StaticResource lblStyle}"/>
 
-                <Button Text="Confirm"
-                       Style="{StaticResource buttonStyle}"/>
-                <Button Text="Cancel" 
-                        Style="{StaticResource buttonStyle}"/>
+				<ListView x:Name="credentialTypes" />
 
-            </StackLayout>
-
-        </StackLayout>
+				<Button Text="Confirm"
+                        Style="{StaticResource buttonStyle}"
+					    Clicked="BtnComfirm_Clicked"/>
+				<Button Text="Cancel" 
+                        Style="{StaticResource buttonStyle}"
+						Clicked="BtnCancel_Clicked"/>
+				
+			</StackLayout>
+		</StackLayout>
     </ContentPage.Content>
 </ContentPage>
