+++ QRCodeAuthMobile/QRCodeAuthMobile/QRCodeAuthMobile/Home.xaml.cs
@@ -3,18 +3,11 @@
  */
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 using QRCodeAuthMobile.Models;
 using QRCodeAuthMobile.Data;
-using ZXing.Net.Mobile.Forms;
-using Newtonsoft.Json;
-using QRCodeAuthMobile.Services;
+
 
 namespace QRCodeAuthMobile
 {
@@ -46,9 +39,9 @@ namespace QRCodeAuthMobile
 			//Application.Current.MainPage = new NavigationPage(new ManageAttendance());
 		}
 
-        private void BtnShareCredentials_Clicked(object sender, EventArgs e)
+        private async void BtnShareCredentials_Clicked(object sender, EventArgs e)
         {
-            scanQRCode();
+			await Navigation.PushAsync(new ConfirmCredentials());
         }
 
         private async void PutUserinSessionState()
@@ -59,77 +52,5 @@ namespace QRCodeAuthMobile
 
             lblWelcome.Text = string.Format("Welcome {0}!", user.FirstName);
         }
-
-        public async void scanQRCode()
-        {
-            var defenition = new
-            {
-                informationCollector = "",
-                department = "",
-                requestedCredentials = new List<CredentialType>() { },
-            };
-
-            //Create a scan page. 
-            var scanPage = new ZXingScannerPage();
-            scanPage.DefaultOverlayShowFlashButton = true;
-            scanPage.HasTorch = true;
-
-            // Navigate to scan page
-            await Navigation.PushModalAsync(scanPage);
-
-            //Event Handler
-            scanPage.OnScanResult += (result) =>
-            {
-                // Pop the page and show the result
-                Device.BeginInvokeOnMainThread(async () =>
-                {
-                    // Stop scanning and dimiss the modal page
-                    scanPage.IsScanning = false;
-                    await Navigation.PopModalAsync();
-
-                    //Save the scanned anonymous object into the obj variable. 
-                    var obj = JsonConvert.DeserializeAnonymousType(result.ToString(), defenition);
-                    string message = "Information Collector : \n" + obj.informationCollector + "\n\nDepartment: \n" + obj.department + "\n\nRequesting Credentials:\n" + "Name\nEmail\nWorkType" + "\n\nWould you like to share your credentials?";
-
-                    var decision = await DisplayAlert("Share Credentials", message, "Yes", "No");
-                    if (decision)
-                    {
-                        // send information to the information collector 
-                        sendRequestedCredentials(obj.requestedCredentials);
-                    }
-                    else
-                    {
-                        await DisplayAlert("Cancelled", "You have chosen not to share your credentials. ", "ok");
-                    }
-                });
-            };
-        }
-
-        public async void sendRequestedCredentials(List<CredentialType> types)
-        {
-            List<Credential> requestedCredentials = new List<Credential>();
-
-            foreach (CredentialType credentialType in types)
-            {
-                Credential cred = await CredentialRepository.GetCredentialByType(credentialType);
-                if (cred != null)
-                {
-                    requestedCredentials.Add(cred);
-                    System.Diagnostics.Debug.WriteLine(cred.CredentialType);
-                }            
-            }
-
-            if(requestedCredentials != null && requestedCredentials.Count > 0)
-            {
-                await DataService.SendRequestedCredentials(requestedCredentials); 
-                await DisplayAlert("Successful!!!", "Your credentials have been shared", "ok");
-            }
-            else
-            {
-                await DisplayAlert("Unsuccessful", "You do not hold one or more of the requested credentials. Therefore, unable to share credentials with information collector. ", "ok");
-            }
-        }
-
-
     }
 }
